feat: add cities data to payload
